Move ParseTable and friends to a new module, jsparagus.parse_table.
